lmp: Make a better "wait_for_update"
The API we were checking doesn't no longer reflects an update
completed until *after* the device reboots. There are probably other
ways in the API to figure this out. However, tailing the log is
gives *much* more live data for the CI job and doesn't require
constantly pinging the server.

Signed-off-by: Andy Doan <andy@foundries.io>