zephyr/jobserv: mark some more tests as flaky
With the switch to nRF52840-DK, we've noticed several tests failing
in sanitycheck.  Most of these are due to specific needs for these
tests.  IE: the USB tests require that the nRF USB be connected during
testing, the led_lp5562 test requires a specific LED be connected via
I2C, etc.

The exception to this is the nffs_fs_api/large test, which needs more
investigation, but is also not working upstream.

Signed-off-by: Michael Scott <mike@foundries.io>