factory-containers: Fix issue with loss of docker-apps
When the server has online keys, we are pulling down the new
targets.json it creates. However, this version loses all the custom
data like docker-apps we have for *each* target. This approach is
actually pretty simple and adds the target by hand every time and
bumps the "version" so that the next push of the targets will be
correct and make the server happy.

Signed-off-by: Andy Doan <andy@foundries.io>