containers: Convert to use "docker manifest"
This tool is now in docker and is the recommended way to build
multi-arch manifests. This change requires us to deal with our old style
MANIFEST_PLATFORMS in a way that's compatible with the new command.

Why do this? The next change that uses "--secrets" produces a new
style container manifest that's not known to the old "manifest-tool".

Signed-off-by: Andy Doan <andy@foundries.io>