lmp/bb-config: add support to build ptest lmp variants via tags
Allow building a special lmp build variant which includes the ptest
packages by default, but via a special tag in order to avoid
contaminating the base builds.

This is done by simply changing the OTA_LITE_TAG variable and appending
-ptest to the lmp-based tags, which is later consumed by the build and
published under a different tag than the one originally intended.

Using a different tag for the ptest-enabled builds allows the user to
easily switch to the ptest tag when performing the OE package testing,
and move back to the base image once that is concluded.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>