lmp: build: fix stripping -b <ref> when hyphen is present
During processing of repo init command we strip potential "-b <ref>"
passed into REPO_INIT_OVERRIDES.  The current logic is:
remove "-b" + <any white space chars> + <any word chars>

If there's a hyphen in the <ref> it thinks this is a new word and
stops removal there.

So if REPO_INIT_OVERRIDES is:
-b home-assistant -m test.xml

We get:
-assistant -m test.xml

To fix this, let's change the removal logic to:
remove "-b" + <any whitespace chars> + <any non-whitespace chars>

This means stop the removal at the next occurance of a whitespace char.

Signed-off-by: Michael Scott <mike@foundries.io>