container preload: Don't do "indent" on script
We generally indent utility scripts to delineate what they are doing,
especially when many scripts are run. In this case its, the script
is basically the CI job. The ident isn't adding much and seems to make
my console output a little choppy (i'm guessing because of long running
commands that aren't printing much and pipe buffering).

Signed-off-by: Andy Doan <andy@foundries.io>