apps: build: drop image label to fix manifest sha
Testing shows that using --label="..." with a different value
each time you run a build will generate a different manifest sha
when pushed to a registry.

This label is used to identify the build a particular image was
created in.  For now, let's drop the label and investigate a
better way to identify images on hub.foundries.io (if at all).

Signed-off-by: Michael Scott <mike@foundries.io>