assemble: determine type of App to preload
If a user doesn't specify an App type to preload then decide on it
taking into account a LmP version. If LmP > 84 then preload `restorable`
Apps by default. If LmP <= 84 or it's not possible to obtain
a LmP version then preload `compose`.

If a user explicitly specifies App type then a LmP version is ignored and
the user defined type of Apps is preloaded.

App type can be defined in
factory-config.yml::containers::preloaded_images::app_type
```
containers:
  preloaded_images:
    enabled: true
    [app_type: [<restorable | compose | '']>]
```

Signed-off-by: Mike Sul <mike.sul@foundries.io>