apps: remove Apps storing and sharing via NFS
Storing and reusing of the "dumped" Apps&images does not bring much
benefit, just decreases a running time for some of the preload/assemble
CI runs (e.g. LmP build's assemble run if Container's assemble run was
executed before).
At the same time, this feature requires usage of NFS, and a required
size of it increases as a number of factories and CI builds grows.
Also, it makes an overall implementation more complex.
Thus, removing this feature makes the overall product simpler and more
reliable.

Signed-off-by: Mike Sul <mike.sul@foundries.io>