lmp: deploy env and logs even in case of failures
It is quite hard to identify build errors (that are not parsing related)
at the moment, since we don't save the environment and the build log is
only available on successful builds.

Fix by making sure the core files are always stored under ${archive}, as
that folder is always uploaded, even on errors.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>