factory_client: Don't throw when no apps and assembling
The first time a target is built on a new branch, it may not have any
apps. The assemble-system-image run fails. Ideally we wouldn't execute
this run, but that's kind of difficult to know ahead of time. This
change is a small compromise that will be the image with no apps.

Signed-off-by: Andy Doan <andy@foundries.io>