lmp: Include IMAGE_LICENSE_CHECHER*DENYLIST variable
After the sensitive review on the class image-license-checker, and the
change of the license tag used on the recipes, the variable name has
been changed. Include the new pair of variable to make it work for the
new class, starting with Kirkstone.

To acomodate the SPDX tags, add asterisks on each license name.

Signed-off-by: Daiane Angolini <daiane.angolini@foundries.io>