apps: Clean up logic around multi arch manifests
We've had a buggy way of creating multi-arch manifests since the
inception of this tool. This predated docker-manifest and the ability of
CI builds to access artifacts from another run. The issue is that we
build each platform concurrently. Therefore, we don't know which run
is going to complete last and be able to create the multi-arch manifest.

This lead to the ugly hack where we basically did:

 docker manifest create && docker manifest push || true

Only the last run will actually publish the manifest. However, you
periodically have infra issues where the manifest push operation
actually fails. You don't *see* that and we move on the publish run
which will eventually fail to pin this multi-arch reference since it
doesn't exist. Then someone in support has to back-track to find the
source of the issue was one of the other CI Runs.

This change fixes the issue by archiving bits of what
docker-manifest-create does for each CI run. The *publish* run can then
combine them all and *publish* the multi-arch manifests. This has a big
improvement from the previous approach where we created the manifest by
*tags*. This change does by sha256 making it impossible to tamper one of
the tags between the build and publish actions.

This work was almost impossible to do without also fixing:

 https://github.com/foundriesio/ci-scripts/issues/148

That was a long-standing issue broken from the move from the old
"mainfest-tool" and to buildx+ docker-manifest. It has a nice
side-effect in that we no longer need to define "MANIFEST_PLATFORMS" as
we've documented:

 https://docs.foundries.io/latest/reference-manual/docker/containers.html#examples

it was left-over junk from manifest-tool.

Signed-off-by: Andy Doan <andy@foundries.io>