lmp: Set uri the apps are inherited from
We do something similar for origUril denoting the platform build a
container build inherited from. This does the same thing and allows us
to see the container build that a platform build inherited from.

In this case we had some old code from ostree apps that hadn't been
removed. I've hijacked that and put in our logic in its place.

Signed-off-by: Andy Doan <andy@foundries.io>