apps: Fix flaky issue with buildkit
We have a customer using a private Google Artifact Registry instance
that's hit buildkit problems. Moving from 0.8.3 to 0.10.3 resolved them
at first. However, as they've added more containers to their factory, we
are seeing a need to create a new context to make them all build
correctly.

This change creates a context for each container and also makes it
easier for us to change what version of buildkit a customer uses.

Signed-off-by: Andy Doan <andy@foundries.io>