apps: Build containers with two labels
We have a problem with factories using a common base image as described:

 https://docs.foundries.io/88/reference-manual/docker/containers.html#id1

Currently, every build is causing the images that extend the base image
to be "rebuilt". I say that in quotes because its still the same image:
every single layer the image is made from remains the same. However, the
docker manifest pointing to the layers does change.

This comes from my misunderstanding of the "FROM" line in the child
containers. Our docs have have basically inherit from
base:${tag}-${arch}. ${tag} is changing every single build. While the
SHA256 it points to does not change, the logic of a Dockerfile to decide
what to build more or less does a string equals comparison of each line
of the Dockerfile. This causes things to rebuild.

We instead need the child containers to have a consistent FROM line:

 ${LATEST}-${arch}

This tag isn't changing but its *sha* will when the base image changes.

This fix moves the logic to evaluate LATEST before we source
pre-build.conf and then have ci publish two tags for each container we
build:

 ${tag}-${arch}    - what we've always had
 ${LATEST}-${arch} - the one we need for base images

NOTE: This allows things to work we still have to fix:

 * factories that use this. Pre-build.conf needs a change to have:
   `_base_img="hub.foundries.io/${FACTORY}/0base:$LATEST-$ARCH"`

 * documentation that fixes the pre-build.conf template

Signed-off-by: Andy Doan <andy@foundries.io>