lmp: Fix logic for detecting the base version
Its a little tricky to pick up the base LmP version. We need to make
sure we've fetched tags from the public LmP and *don't* look at tags
that have been created by the user.

The old version was also getting the *latest* tag which is currently v89
but that's probably not the actual LmP version the customer is building
with. Git doesn't have a way to distinguish where tags come from, so the
easiest thing to do is just delete all the tags and then pull from our
official tags.

This logic is also getting a bit complex, so I've moved it to a
dedicated function.

Signed-off-by: Andy Doan <andy@foundries.io>