set_base_lmp_version: fix LMP_VERSION detection
The version need to be extracted from the upstream lmp-manifest
otherwise it can come from the user repo that can have invalid
tags for this propose or even not having any tag.

Also change it and use abbrev=0 to get only the tag name and in
this way we can use this as well for the sstate cache.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>