lmp/bb-config: the HTTP sstate mirror is the default now in LmP
The default value of the SSTATE_CACHE_MIRROR is the mount path of the
shared NFS sstate cache mounted on the SDK build container.
So when the folder in SSTATE_CACHE_MIRROR exist, the NFS is chosen.

This happens when the worker have access to the NFS shared cache.
On the "-nocache" workers, the HTTP mirror will be used because the
shared NFS is not available on the worker and consequently
is not mounted on the SDK build container.

The variable SSTATE_CACHE_MIRROR can also be changed on the factory .yml.
So for example SSTATE_CACHE_MIRROR="" will disable the NFS mirror,
being used in this case the default one HTTP.
This add the possibility to use the HTTP sstate cache mirrors in every
factory if it is required.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>