lmp/bb-config: don't use the mirror only for the first build
This currrentlly disable the SSTATE_MIRRORS after the first build because
the local sstate cache SSTATE_DIR folder ../sstate-cache is only populated
on the first build.

we should always use SSTATE_MIRRORS so the factories can take advantage
of the public cache when updating from v(N-1) to v(N).

Bitbake will check for sstate cache in the following order:
 1 - SSTATE_DIR
 2 - SSTATE_MIRRORS

So it will only search the mirror if the file not found on the local cache ../sstate-cache,
this is an insignificant overhead searching the SSTATE_MIRRORS because most of the files
are on the local SSTATE_DIR.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>