apps: Create a helper for docker-login
This consolidates the docker-login logic to a single helper.

This also changes the container build process only call docker-login
once against hub.fio which is all that's needed. Calling multiple times
has no effect. We are also long past the time you could run this script
with "noauth", so I simplified the logic

Signed-off-by: Andy Doan <andy@foundries.io>