assemble: Preload apps to docker store unconditionally
Preload Apps to the dockerd store unconditionally even if
the restorable apps are preloaded to the skopeo/OCI store.

The reason for doing so is to optimize the timing of
the app's "early" startup by eliminating the need to copy the app's
data (layers/images) from the skopeo store to the dockerd store during
the early startup process. The copying process can be time-consuming on
resource-constrained devices, as it involves untarring and
decompression of data.

Signed-off-by: Mike <mike.sul@foundries.io>