lmp/bb-config: drop rm_work inherit
If the CI worker have enough space to accommodate everything until the end this can be a good improvement.
Instead of removing what we no longer need at the end of each bitbake task we
can drop everything at the end, the time reduced it's not much but it's about 5 minutes.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>