apps: Make the sbom logic pull pinned factory images
A user may refer to a factory image with a pinned/sha256 digest in
`docker-compose.yml`. It implies that the given image is not built in
the current CI run, hence the image is not analyzed by the SBOM logic
for images built in the current CI run.
This change makes the SBOM logic for non-factory images pull
the pinned factory images, thus all App's images are SBOMed.

Signed-off-by: Mike <mike.sul@foundries.io>