deltas: Make sure pull items are unique
Use `set` as a container for download URLs and commit hashes to ensure
that ostree repo/commits are pulled ones.
The static delta handler at ota-lite does not take into account the
multitag case that leads to two or more targets with different name but
the same reference to ostree repo/commit.
It's easier to remove duplications in the script than in ota-lite
handler, hence this change does so.

Signed-off-by: Mike <mike.sul@foundries.io>