lmp: add factory sstate-cache mirror
In this new layout bitbake will start with a clean sstate-cache folder but
with the help of the persistent mirror stored on each factory.
When the build finish we rsync back to the mirror the new cache files created
in this build.

We still keep using the https public mirror but we will prioritize the local one
of the factory.

It still possible to change the sstate-cache mirror in the CI build because
the mirror is assinged with a weak default ?= so it can be change with =.
For example it can be disabled with SSTATE_MIRRORS = ""

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>