preload: Fetch and store manifest of app images
For some reason `skopeo` doesn't store an image root, index or manifest
in a store blob directory. Instead it creates a custom manifest file in
the image directory that refers to the image manifest stored in the blob
one. It is OK if an image root is actually manifest, but it doesn't OK if
an image root in an index - missing node/blob of the image merkle tree.
As result, there is no way to check image integrity on devices with
preloaded apps/images.
The new compose app management utility does check integrity of an app's
merkle tree starting from its root node. Therefore, to let it work
properly we need to store an index of each app image and
the given change does it.

Signed-off-by: Mike Sul <mike.sul@foundries.io>