helpers: Hack around MEDs/oe-builtin issue
oe-builtin logic is seeing a CI error like:
| Traceback (most recent call last):
|   File "/script-repo/preload_apps.py", line 14, in <module>
|     from apps.docker_registry_client import ThirdPartyRegistry
|   File "/script-repo/apps/docker_registry_client.py", line 18, in <module>
|     class DockerRegistryClient:
|   File "/script-repo/apps/docker_registry_client.py", line 19, in DockerRegistryClient
|     DefaultRegistryHost = 'hub.' + fio_dnsbase()
|   File "/script-repo/helpers.py", line 123, in fio_dnsbase
|     parts = urlparse(os.environ["H_RUN_URL"])
|   File "/usr/lib/python3.8/os.py", line 675, in __getitem__
|     raise KeyError(key) from None
| KeyError: 'H_RUN_URL'

This is because bitbake isolates the environment that it runs in. There
are a couple ways to pass H_RUN_URL in, but they cause a big cache
invalidation.

The workaround provided here works as long as we don't have a customer
on MEDs that does oe-builtin.

Signed-off-by: Andy Doan <andy@foundries.io>