apps: Enable OCI image format
Add ability to enable building images in the OCI format for a factory.
By default, the docker format is applied as it is now. To enable the OCI
format the `ENABLE_OCI_IMAGES` should be set to "1".

The Docker format is gradually becoming obsolete, with most advancements
now focused on the OCI image format. This change enables us to perform
end-to-end testing of OCI image format support.
After that, we can offer our customers the option to transition to the new
format and take advantage of its enhanced capabilities, such as
switching to zstd compression.

Signed-off-by: Mike Sul <mike.sul@foundries.io>