lmp/jobserv: remove the TI machines on the scarthgap-next
For the v95 scarthgap based the meta-ti layer are not compatible with the
scarthgap-next oe-core fork we currently have but it builds without it well.
So I will remove the TI layer and machines from CI job because it brings
no value to continue to exercise this than always fails.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>