lmp/bb-build: clarify the bitbake notification
- add notification about the end
- change the finish notification to post processing

When there are errors during the bitbake build it is always necessary
to perform some additional post-processing steps such as:

== Run bitbake (finish)
== Run bitbake (rsync factory sstate-cache mirror)

Script completed with error(s)

Only after this last step is reported that there was an error.
This leads the user to think that the error was in the last step,
in this case the rsync, which is not true.

Example:

Summary: 1 task failed:
  /srv/oe/build/conf/../../layers/meta-subscriber-overrides/recipes-samples/images/lmp-factory-image.bb:do_image_garagecheck
Summary: There were 4 WARNING messages.
Summary: There was 1 ERROR message, returning a non-zero exit code.
== 2024-10-10 12:00:20 Run bitbake (finish)
bootchart written to '/archive/bitbake_buildchart.svg'
== 2024-10-10 12:00:24 Run bitbake (rsync factory sstate-cache mirror)
Script completed with error(s)
== 2024-10-10 12:01:32.828429: Finding artifacts to upload
Uploading 14 items 53779607 bytes
== 2024-10-10 12:01:36.053422: Runner has completed
          ________
          |  o  o|           Thumbs Down
          |___/\_|________
          |       _____|__|
          |      |     ||
          |______|
          | |  | |
          | |  | |
          |_|  |_|

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>