lmp/bb-build: only use the sstate-cache mirror when it is necessary
Our build can be divided into two steps:
1 - build only using the sstate-cache
2 - build everything else missing

On the step [1] we run bitbake for setscene only tasks
and obvious this one is the only one that uses and reads the
sstate-cache mirror.

The step [2] will have 0% hit of cache available because everything
has already been accelerated in the step [1]. In the step [2]
everything will be build from scratch and the local sstate-cache
will be populated with the new artifacts.

This change will improve and significantly reduce the load generated on
the sstate-mirror http server by our CI builds.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>