lmp/build: add support for collecting duplicated sources
On multiconfig machines the sources and the license can be duplicated.
We have two places with the source-code and licenses on the ti bsp [1].
If it was the case we will deploy only one package, we will search for
the first package license path and use that.

We can also escape and jump this source-code collecting step when we not have any IMAGE_MANIFESTS (i.e no image produced).

[1] https://github.com/foundriesio/meta-lmp/blob/7657b41046ff745ce5fe61a64f615623ebb098e0/meta-lmp-bsp/conf/machine/include/lmp-machine-custom.inc#L457-L459

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>